Bigarren sprint2 (Pull request aldaketak egindak
