+++ Erronka3/src/Kontroladorea/Kontroladorea.java
@@ -4,9 +4,10 @@ import DAO.BezeroaDAO;
 import DAO.SaioaDAO;
 import DAO.ZinemakDAO;
 import Modelo.Bezeroa;
+import Modelo.Erosketa;
 import Modelo.Saioa;
 import Modelo.Zinema;
-import Vista.OngiEtorri;
+import Vista.OngiEtorriV;
 
 public class Kontroladorea {
 
@@ -15,8 +16,9 @@ public class Kontroladorea {
         Zinema[] zinemakList = zinemakDAO.getAllZinemak();
         BezeroaDAO bezeroaDAO = new BezeroaDAO();
         Bezeroa[] bezeroak = bezeroaDAO.getAllBezeroak();
+        Erosketa erosketak = new Erosketa();
         try {
-            OngiEtorri frame = new OngiEtorri(zinemakList,bezeroak);
+            OngiEtorriV frame = new OngiEtorriV(zinemakList,bezeroak, erosketak);
             frame.setVisible(true);
         } catch (Exception e) {
             e.printStackTrace();
