+++ Erronka3/src/Modelo/Karteldegia.java
@@ -18,13 +18,6 @@ public class Karteldegia {
 		this.filmak = filmak;
 	}
 
-	@Override
-	public int hashCode() {
-		final int prime = 31;
-		int result = 1;
-		result = prime * result + Arrays.hashCode(filmak);
-		return result;
-	}
 
 	@Override
 	public boolean equals(Object obj) {
